Updated the modal/dialog
